Learned Pattern stuff from invalid
added 2 fake 1 damage ammos that fire in step with the main ammo, and all add trail decay to the turret so it looks great.  Increased reload to be standardized with trident.  Probably should touch other reloads of the big turrets to be standardized with trident but those are being tested and thus i wont touch them yet.