Tigers did not, in fact, need a x8 shield mod.
Reverting to x3, fixing some of the heat instead of dropping other damages.

40 delay, 350 reload, means it reaches exactly max heat after the 4th burst of the 3rd salvo, and not in-between bursts, thus reloading when overheated.  0.55 heat cap means it can fire off 1 full salvo of 4 bursts before overheating again and not overheating mid salvo.

loads a 16 round clip, meaning that a full salvo means 4 bursts of 4 bullets each

T1s are in! Reducing range so they are not the same range of a T2 and better at PD than a much bigger T2 laser.

Looping particles must perish.