attempt to improve siege formula
found an issue of discontinuity in the siege math, this should fix it.  Set min siege time from 15 -> 20 for now.
The issue was any time spent in siege between 8 seconds and 15 seconds was worth an incorrect 15 seconds of cooldown, and the 16th second was worth a correct 32 seconds of cooldown.

Now there is a smooth transition, the first 10 seconds are worth 20 seconds of cooldown, and the 11th second is 22s of cooldown, the 12th second is 24s of cooldown, etc.

Tested and confirmed to work in single player.